Fix uptime parser bug, thermal culture bug, and address code review findings
Fix UptimeQuery dropping minutes from H:MM format ("1:31, 0 users" yielded
1h0m instead of 1h31m) and hours from "days, H:MM" format by searching for
the "user" keyword to delimit the uptime span.

Fix ThermalQuery summary using current culture decimal separator by wrapping
with FormattableString.Invariant.

Also: add IsTestProject to test csproj, label CPU stat arithmetic fields,
add PacketsOut comment, document IsAllDigits vacuous truth, update tag-release
script with alpha/beta/rc/stable pre-release support.

+++ src/ShieldCommander.Core.Tests/Queries/ThermalQueryTests.cs
@@ -44,9 +44,10 @@ public class ThermalQueryTests
 
         await Assert.That(result.Zones).Count().IsEqualTo(5);
         await Assert.That(result.Zones[0].Name).IsEqualTo("CPU0");
-        await Assert.That((float)result.Zones[0].Value).IsEqualTo(56.000004f);
+        // Value passes through float.TryParse, so compare at float precision
+        await Assert.That(result.Zones[0].Value).IsEqualTo(56.000004f);
         await Assert.That(result.Zones[4].Name).IsEqualTo("GPU");
-        await Assert.That((float)result.Zones[4].Value).IsEqualTo(55.500004f);
+        await Assert.That(result.Zones[4].Value).IsEqualTo(55.500004f);
     }
 
     [Test]
